+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v2.9.22
 
 * minor typing clarifications
-
+* added `length_factory` to `hint_with_length`